move Glyph to results.py
