+++ flogin/jsonrpc/__init__.py
@@ -1,5 +1,4 @@
 from .client import *
 from .errors import *
-from .glyph import *
 from .responses import *
 from .results import *
